refactored angular modules to modularize
