+++ angular_app/about/about.module.js
@@ -1,0 +1,2 @@
+// Define the `phoneList` module
+angular.module('about', []);
