+++ index.html
@@ -21,6 +21,8 @@
 
   <script src="node_modules/angular/angular.js"></script>
   <script src="angular_app/app.module.js"></script>
+  <script src="angular_app/about/about.module.js"></script>
+  <script src="angular_app/about/about.component.js"></script>
 
 </head>
 <body ng-app="ddsDesktop">
