+++ sections/about.html
@@ -20,7 +20,7 @@
         </div>
         <section class="about-section about-code">
           <h2>About the Software Agent</h2>
-          <p>The Duke Data Service is a platform for reproducible research. There are several resources online for those
+          <p>Duke Data Service is a platform for reproducible research. There are several resources online for those
             interested in learning more. The backend is available on <a href="https://github.com/Duke-Translational-Bioinformatics/duke-data-service">Github<span class="u-visible-to-screen-reader">(opens in new window)</span></a>
             and is open-sourced. In order to use this application, you will first need to setup a software agent on the
             <a href="https://dataservice.duke.edu">duke data service portal<span class="u-visible-to-screen-reader">(opens in new window)</span></a>.
